+++ NEWS.md
@@ -5,6 +5,7 @@
 * fix: Removed parameters from RandomForestSRC learners that weren't used + optimized tests
 * fix: Remove `discrete` parameter from `surv.parametric`, so that it is impossible to return `distr6::VectorDistribution` survival predictions (softly deprecated in `mlr3proba@v0.8.1`)
 * feat: Added learner `LearnerSurvBlockForest`.
+* feat: Added learner `LearnerClassifAdabag`.
 * feat: Added learner `LearnerClassifFastai`.
 * feat: Added learners `LearnerRegrPenalized` and `LearnerClassifPenalized`.
 * feat: Added learners `LearnerClassifBst` and `LearnerRegrBst`.